move metadata related stuff to module
